Update all Form Requests Update Controllers to redirect to model index Update to show validation for all trumbowyg textareas Added Clear button for notifications for UX General Bugfixes and Stability Improvements

+++ app/Http/Controllers/QuoteItemController.php
@@ -81,6 +81,10 @@ class QuoteItemController extends Controller
      */
     public function destroy(QuoteItem $quoteItem)
     {
-        $quoteItem->delete();
+        $quote = $quoteItem->quote;
+        if($quote->items->count() != 1)
+        {
+            $quoteItem->delete();
+        }
     }
 }
